projects/pam-private-asset-management/pam-backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
passlib
python-jose
python-multipart

